# back_end/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Manual do Skatista",
      "version": "1.0.0",
      "description": "Documentação da API de usuários (CRUD + login)"
    },
    "servers": [
      {
        "url": "http://localhost:3001"
      }
    ],
    "paths": {
      "/usuarios": {
        "post": {
          "summary": "Cadastrar um novo usuário",
          "tags": ["Usuarios"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuario cadastrado"
            }
          }
        },
        "get": {
          "summary": "Listar todos os usuários",
          "tags": ["Usuarios"],
          "responses": {
            "200": {
              "description": "Lista de usuários",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{id}": {
        "put": {
          "summary": "Atualizar um usuário",
          "tags": ["Usuarios"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuario atualizado"
            }
          }
        },
        "delete": {
          "summary": "Deletar um usuário",
          "tags": ["Usuarios"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario deletado"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Usuario": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "nome": {
              "type": "string",
              "example": "João Silva"
            },
            "email": {
              "type": "string",
              "example": "joao@email.com"
            },
            "senha": {
              "type": "string",
              "example": "123456"
            }
          }
        }
      }
    }
  }